kleaf: temporary workaround for target sibling calculation.
For example, if a dependency is //package/to/foo:foo, the label
will be reduced to //package/to/foo by buildifier / buildozer.
However, appending suffixes to this string is incorrect.

Work around this by properly parsing the label, then appending the suffix.

The proper fix should be to use transitions for siblings (b/237706175)

Test: TH
Bug: 237706175
Change-Id: Iec669ece96bc25bd7cbfb7f6293ad104a6f36f54