kleaf: kernel_build_abi targets use transitions
... instead of the awkward siblings targets.

Instead of defining name_with_vmlinux and name_notrim
targets for kernel_build and kernel_module's, use transitions
so that the ABI targets (e.g. when extracting symbols
in extracted_symbols) depend on the proper variant of
the kernel_build / kernel_module with a slight change in
the configuration.

The only behavioral change in this CL is that the GKI modules
used to be looked up from //common:kernel_aarch64; but now they
are looked up from the notrim variant. Trimming only affects
vmlinux, not modules, so this behavioral change in the analysis
phase should not have effect on the final result.

Bug: 237706175
Test: TH
Test: manual on Pixel

Change-Id: Ie352b15bc960d810e53a5ea2446afdeffe67cd54