kleaf: Export UAPI headers as native cc_library
There's currently no way for userspace programs building with
Bazel to ingest the kernel UAPI headers. Kleaf generates a tarball
of the headers, but a Bazel cc_binary rule cannot use that as an
input.

Define a native header-only cc_library which userspace programs can
use to build with the kernel's UAPI headers. For example:

cc_binary(
    name = "foo",
    srcs = ["foo.c"],
    deps = ["//common:kernel_aarch64_uapi_header_library"],
)

Bug: 258176225
Change-Id: I06cc582992460e575d50cde501c47c82566291c3
Signed-off-by: John Moon <quic_johmoo@quicinc.com>