kleaf: Add rule to unpack UAPI headers
There's currently no way for userspace programs building with
Bazel to ingest the kernel UAPI headers. Kleaf generates a tarball
of the headers, but a Bazel cc_binary rule cannot use that as an
input.

Define a rule to unpack the UAPI headers so that userspace builds
can use them. For example:

kernel_unarchived_uapi_headers(
    name = "kernel_aarch64_unarchived_uapi_headers",
    kernel_uapi_headers = "//common:kernel_aarch64_uapi_headers",
)

cc_library(
    name = "uapi_header_lib",
    hdrs = [":kernel_aarch64_unarchived_uapi_headers"],
    includes = "kernel_aarch64_unarchived_uapi_headers",
)

cc_binary(
    name = "foo",
    srcs = ["foo.c"],
    deps = [":uapi_header_lib"],
)

Bug: 258176225
Change-Id: I06cc582992460e575d50cde501c47c82566291c3
Signed-off-by: John Moon <quic_johmoo@quicinc.com>