kleaf: Assert assumptions on ddk_module depending on submodules.
If a makefiles target has submodules makefiles, it must not
declare any inputs, outputs or copts.

If a makefiles target does not have submodules, it must
declare an output.

This ensures that we only allow two cases:
- A single ddk_module with out specified
- A ddk_module with some ddk_submodule, and the
  top-level ddk_module must be empty.

Also enforce that submodules must be in the same package.
The logic to merge Kbuild files expects that ext_mod is
the same for both the submodule and the high-level module.

Test: TH
Bug: 257319565
Change-Id: I2422f17984d6f72ed897e698bac63209f0caac79