kleaf: Introduce ddk_submodule
ddk_submodule's allows us to build multiple ko
files within the same ddk_module. This is so that the
dependencies between these ko files do not need to
be specified.

This is especially useful when you have multiple ko
files for the same subsystem, where the drivers are
dependent on each other, and maintaining the list of
dependencies for each driver is painful.

Bug: 257319565
Test: see next CL

Change-Id: I85322abeeaa59f30d39e9ee2cdbefc253e9cd45e