kleaf: shorten OUT_DIR_SUFFIX.
This makes the path shorter and easier for the eyes.
To preserve debugging information, create a
kleaf_config_tags.txt under OUT_DIR to know the
config tags.

Caveat: We are using a 32-bit hash function similar
to Java's String.hashCode(), which can cause hash
collisions. However, we usually don't have too many
variations of the string, so let's stick with this
now.

Test: pass
Bug: 260955234
Change-Id: I84f9205b18f315aee7d580570f7e8d62fbfbb306