build_utils.sh: unconditionally strip modules that go on the device
Currently we are already doing this based on the DO_NOT_STRIP_MODULES
flag. When the flag is not set, the stripping happens by the kernel via
INSTALL_MOD_STRIP=1 flag. To simplify the logic, let's just
unconditionally strip in create_modules_staging(). This is addressing
the current issue where kleaf doesn't strip these modules when
`strip_modules = False`.

Bug: 261513123
Change-Id: Id7c0d2229a3e79afb7881ca58421cfa61debf320
Signed-off-by: Will McVicker <willmcvicker@google.com>