Rework kernel_kythe
Instead of relying on interceptor compile_commands,
rely on compile_commands from Kbuild instead.

Also fix the long-broken kythe target. Note that for
a regular kernel build, Kbuild does the following:

- cd ${OUT_DIR}
- generate intermediate files into ${OUT_DIR}
- source files are referenced via absolute paths

Hence, in the generated compile_commands.json from Kbuild:

- Relative paths refers to files in OUT_DIR
- Absolute paths appears to be outside of OUT_DIR

The best way for this to work with Kythe is to combine
OUT_DIR into ROOT_DIR/KERNEL_DIR, so the source files
appears to be

Bug: 201801372

Change-Id: I15878b3d737fbacca4af0903e5fe54e2decfd6d7