Revert "download_from_ci: add 16k experiment"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4f6e930a-124c-48d6-826f-d0b980e15795). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Idb5069b3a9fcc37091a5cae9fd1ad5412dbb14f1