kleaf: Add support to get copy of OUT_DIR unconditionally
The `OUT_DIR` is hided inside kleaf, unlike build.sh, we cannot
get the path of `OUT_DIR` before build system starting.
Unfortunately, some of automatically
debugging and analyzing tools which need static `OUT_DIR`
path would be broken.

`--get_kernel_output` is for compatibility with Linux build and
build.sh build which get `O`, a.k.a. `OUT_DIR` in kleaf,
easily and get everything unconditionally.

This is only for debugging or analyzing which would NOT affect
any sandbox or caching mechanism (e.g. config=local) in Bazel.

When the flag `--get_kernel_output` is set, the
`OUT_DIR` would be rsynced to bazel output.

Bug: 261145969
Change-Id: Ic5d6a2137c6d932b804274231b46cdee0c440459
Signed-off-by: Lecopzer Chen <lecopzer.chen@mediatek.com>