Revert "kleaf: Rework kernel_kythe"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ef2df5b3-caba-4c25-8e38-c4c8f88277c9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ifbdc06e344816ef5afeb207377d250666c2a093a