kleaf: modules_prepare: tar --mtime
for a more reproducible and deterministic output
to reduce rebuilds in Bazel.

Because kernel_module's depend on the modules_prepare
target, if it is rebuilt and the artifacts aren't
deterministic, all kernel modules will be rebuilt. Making
the artifacts deterministic reduces rebuilding of kernel
modules for incremental builds when a core kernel file
is changed.

Bug: 254357038

Change-Id: I46336cb0f36ff3553d881c820db39de38f48b5b9