kleaf: Move env_info from KernelEnvAttrInfo to KernelBuildOriginalEnvInfo.
env_info is not an attribute of kernel_env, so it
does not belong to KernelEnvAttrInfo. Now that
kernel_config exposes the original kernel_env[KernelEnvInfo] from
kernel_env in KernelConfigEnvInfo, kernel_build
can now provide a separate provider, KernelBuildOriginalEnvInfo,
instead of KernelEnvAttrInfo, to expose the kernel_env[KernelEnvInfo].

This is a refactoring change and there should not be any behavioral
changes.

Bug: 201801372
Test: manually build //common:kernel_aarch64_kythe

Change-Id: I1a2262849421a0b2d68b07e669eace4dbe409700