kleaf: Also cache out_dir for kernel_config.
The change moves the OUT_DIR under out/cache when
building kernel_config, similar to how we do
for kernel_build.

This is not intended as an optimization; rather, this
change ensures that OUT_DIR outside of the sandbox is
never touched, because it may cause conflicts for various
rules. That is, we do not want to write to:

  out/bazel/output_user_root/<hash>/execroot/__main__/out${OUT_DIR_SUFFIX}

Test: TH
Test: manual
Bug: 254357038
Bug: 256715367
Change-Id: I4d117430577603ebd3b21cd0cb538a254acbf3dc