Optionally provide Rust toolchain to the kernel build
If RUSTC_VERSION is defined in common/build.common.constants, then
the kernel kleaf build will use that information to find rustc,
cargo and bindgen. With these tools, the kernel build will set
CONFIG_RUST_IS_AVAILABLE=y which then allows CONFIG_RUST=y to
be set.

Note some (coming) changes to the Rust toolchain prebuilts are
also needed. Those will be addressed in a separate change, but
this change is safe to use with or without those changes.

Bug: 270672379
Change-Id: Ib221d94984f719d3d7d257a25856c44b02b57b1b