build.sh: Remove stale devicetree binaries
Currently, the devicetree build artifacts (e.g. .dtbo and .dtb files)
are output to a directory, and they are later used by mkdtimg to create
the dtbo image. When SKIP_MRPROPER=1 is used, the devicetree artifacts
are not removed.

This is problematic in cases where the devicetree makefile has been
modified to exclude a dtbo that was previously compiled, as the dtbo
file will still be in the output directory, and will be included in the
dtbo image. Thus, remove all devicetree artifacts from the output on
every compilation.

Bug: 264602319
Change-Id: I9ed961dc6729cf66f8b7f158ea3298e29c232a22
Signed-off-by: Isaac J. Manjarres <isaacmanjarres@google.com>