Revert "kleaf: kernel_config keeps fixdep"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/98f952df-8096-49a2-9f48-f795ac958499). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4a3d540919f1d47381920811448831f1023ee4a4