Revert "kleaf: kernel_build use out_dir_suffix for its own."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/98f952df-8096-49a2-9f48-f795ac958499). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7b814521ff65e4464d391612b3fd8da3d68794a1