Revert "kleaf: Also cache out_dir for kernel_config."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/98f952df-8096-49a2-9f48-f795ac958499). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2aa87607134a4658d45a45319c9b9f99ea2a7d2d