Revert "kleaf: kernel_build avoid writing to OUT_DIR under execroot."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/98f952df-8096-49a2-9f48-f795ac958499). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If99f5f3ce457198439ed5d3661d62fb577a57b69