build_utils.sh: Use modules.order for EXT_MODULES_MAKEFILE too
If drivers are written correctly, then it shouldn't matter the order
that they are loaded in. So in order to support using the modules.order
files with EXT_MODULES_MAKEFILE, we need to drop the dependency on
EXT_MODULES. This allows us to still retain loading modules in the
Makefile build order within a given external module project.

Note: this patch makes the order of loading external modules
indeterministic since we are not going to load based on the results of:

  find ${dest_dir}/extra -name modules.order.* -print

Bug: 210713925
Change-Id: I5b92b233bbd75a29ba137bf6e2658d3cab55964b
Signed-off-by: Will McVicker <willmcvicker@google.com>