kleaf: kernel_abi_dist only depends on kernel_build (with_vmlinux). (2nd attempt)
For example, kernel_aarch64_abi_dist only depends on
kernel_aarch64 (with_vmlinux). This is so that we aren't
building kernel_aarch64 twice, once for ABI artifacts,
the other for dist artifacts.

This option is not enabled by default. The
kernel_abi_dist.kernel_build_add_vmlinux attribute
controls whether this is enabled. In the future, when
devices are fixed properly, we'll enable this by default.

Bug: 264710236
Test: manual

Change-Id: I2542aa4a7a889dcfe635be885a4b2021235902ec