kleaf: refactor: KernelConfigEnvInfo -> KernelEnvAndOutputsInfo.
The new provider contains more appropriate information
for environment setup that restores output to OUT_DIR.
It contains a few things:

- A setup script generator FUNCTION;
- depset of inputs / tools for the fixed scripts in the function

The setup script generator is invoked by the dependent
targets which may or may not be executed in the sandbox.
For those dependent targets not in the sandbox, they must
provide a command (cache_dir_step.cmd, from cache_dir.bzl)
that redirects OUT_DIR properly to the generator function.

The generator function returns a script that does the following:
- Set up environment variables
- Invoke the script provided by dependent targets that
  may or may not redirects OUT_DIR
- Restore outputs to OUT_DIR.

As part of this change, all KernelEnvInfo.dependencies (list[File])
are migrated into two depsets, inputs and tools, because:

- inputs are compiled to be executed on the target platform
  (the Android device), and tools are compiled to be executed on
  the execution platform (the build machine)
- depsets are more efficient than list[File] for large lists.

Bug: 247624301
Bug: 263385781
Bug: 254357534

Change-Id: I295c30a280bf44798a49e322566ab26c4b2248b8