Revert "kleaf: Delete unused vmlinux_btf.bzl."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0c37f4eb-609e-4da5-9474-f6df283e19a9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I809230f5f0660bb1fbd46e2926b6ae5f94cd4f69