kleaf: DDK: support kbuild_options (WIP).
Implementation detail: Instead of actually setting
KBUILD_OPTIONS in the top level makefile, this implementation
sets the individual variables in each generated Kbuild files.

This is so that it works with ddk_submodule; that is,
a kbuild_options set by a ddk_submodule does not interfere
with another ddk_submodule in the same ddk_module.

Note that this won't set local_defines nor copts, so you
can't read the values in the source files. This is primarily
useful for selecting files in conditional_srcs.

Test: pending unit tests
Bug: 264293217
Change-Id: I57359440f9835e02c6b3553e73dcc4472c13e969