kleaf: Filter non mixed tree files for kernel_filegroup
When we specify a kernel_build target in kernel_filegroup.srcs,
the old behavior is to get all files from DefaultInfo and
put it into KernelBuildMixedTreeInfo. However, some files
(e.g. *_modules, *_checked, *_gcno etc.) are not needed
for mix builds.

Hence, if KernelBuildMixedTreeInfo is detected in a target
in kernel_filegroup.srcs, use it to form KernelBuildMixedTreeInfo
instead.

This fixes a bug that, because the _gcno directory is included in
DefaultInfo, kernel_aarch64_download_or_build exposes it in
KernelBuildMixedTreeInfo incorrectly (while kernel_aarch64 correctly
omit it). Then in device builds, creating KBUILD_MIXED_TREE incorrectly
symlinks into the sandbox (because in the sandbox, TreeArtifact's
are a real directory, but the contents are symlinked).

Test: TH
Bug: 267818415

Change-Id: Id27b048999c69842b86d67cc91417f6c56c477aa