Revert "kleaf: abi: Add target to update stg file"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/38b3a80f-4f7e-4ff4-97eb-21e74c2dc806). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I536443502ab37caeab8a80b0d75c8dbcb2ee5fc9