ABI: Add script to verify trimmed ksymtab
The utilization of abi/compare_to_symbol_list is to be substituted with
abi/verify_ksymtab.py. The former performs two functions.

Firstly, it ensures that all symbols specified in the vendor symbol list
are exported in ksymtab, which is mandatory for enforcing KMI strict
mode. The latter, i.e., verify_ksymtab.py, replaces this primary
functionality.

Secondly, it verifies that all symbols used by the vendor modules are
listed in the vendor symbol lists to avoid trimming. This is imperative
to ensure that the symbol list is accurate and prevent runtime module
loading issues. If there is no symbol list, the runtime check does not
enforce this verification and assumes that there is no ABI in place to
support targets like x86_64. This check is being replaced by
aosp/2435512.

Bug: 269346251
Test: http://b/269346251#comment10
Test: TH
Change-Id: I5c9dcb39e33211ec8d221ba55234ac813de3c6e6
Signed-off-by: Ramji Jiyani <ramjiyani@google.com>