ABI: Add script to verify trimmed ksymtab
Converts the abi/compare_to_symbol_list to a python
script. It checks missing symbol from the ksymtab as
in the original compare_to_symbol_list script whenever
KMI strict mode is enabled.

It removes the check for the missing symbol from the
symbol list. Since this check is required to perform
as long as there is a symbol list to enfoce correctness
of the symbol list irrespective of the KMI strict mode.
It is performed by check_buildtime_symbol_protection.py

Bug: 269346251
Test: http://b/269346251#comment10
Test: TH
Change-Id: I5c9dcb39e33211ec8d221ba55234ac813de3c6e6
Signed-off-by: Ramji Jiyani <ramjiyani@google.com>