build.sh: Copy GKI artifacts before building the device kernel
We need to copy the GKI artifacts to the device DIST_DIR before building
the vendor device kernel so that we don't overwrite any vendor modules
with the same name as the GKI modules. This allows vendors or override
GKI modules.

Note: since build.sh is being deprecated for android14+ kernels, we
don't need to worry about non-overridable GKI modules.

Change-Id: I4982a7ff795da328d08a422b757114899f3f7410
Signed-off-by: Will McVicker <willmcvicker@google.com>