Revert "Revert "kleaf: update comment for test_config_uapi_heade..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f0788fd7-5ad7-44a1-a3c8-2e1c417ef959). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I44a8e410a6dd8b6eab30e3925873276219737eac