kleaf: CONFIG_LOCALVERSION -> $OUT_DIR/localversion
Instead of having CONFIG_LOCALVERSION to store the localversion
for vmlinux, rely on a second mechanism in the setlocalversion
script: $OUT_DIR/localversion* files.

This ensures that CONFIG_LOCALVERSION is unset, so that
the SCM version of the core kernel (vmlinux) does not pollute
the SCM version of external modules when external modules
move away from .scmversion file.

Bug: 272828027
Test: integration test

Change-Id: I738e9d0120fd3e5c9426dc650514fbf41e967439